// src/types/questionarry-list.ts
import { z } from "zod";

const streetTypeSchema = z.object({
  acronym: z.string(),
  name: z.string(),
});

const streetSchema = z.object({
  id: z.number(),
  name: z.string(),
  streetType: streetTypeSchema,
});

const neighborhoodSchema = z.object({
  id: z.number(),
  name: z.string(),
});

const federateUnitSchema = z.object({
  acronym: z.string(),
  name: z.string(),
});

const citySchema = z.object({
  id: z.number(),
  name: z.string(),
  federateUnit: federateUnitSchema,
});

const addressSchema = z.object({
  id: z.number(),
  cep: z.string(),
  street: streetSchema,
  neighborhood: neighborhoodSchema,
  city: citySchema,
});

const specificAddressSchema = z.object({
  number: z.string(),
  complement: z.string(),
  address: addressSchema,
});

const dddSchema = z.object({
  ddd: z.number(),
});

const ddiSchema = z.object({
  ddi: z.number(),
});

const phoneSchema = z.object({
  number: z.string(),
  ddd: dddSchema,
  ddi: ddiSchema,
});

const cpfSchema = z.object({
  cpf: z.string(),
});

const genderSchema = z.object({
  acronym: z.string(),
  description: z.string(),
});

const studentSchema = z.object({
  id: z.number(),
  name: z.string(),
  specificAddress: specificAddressSchema,
  phoneList: z.array(phoneSchema),
  mainEmail: z.string(),
  recoveryEmail: z.string(),
  cpf: cpfSchema,
  gender: genderSchema,
  birthDate: z.string(),
});

const questionAlternativeSchema = z.object({
  id: z.number(),
  statement: z.string(),
  isCorrect: z.boolean(),
});

const questionDifficultyLevelSchema = z.object({
  id: z.number(),
  name: z.string(),
  description: z.string(),
});

const subjectSchema = z.object({
  id: z.number(),
  name: z.string(),
  description: z.string(),
});

const subjectTopicSchema = z.object({
  id: z.number(),
  name: z.string(),
  description: z.string(),
  subject: subjectSchema,
});

const questionSchema = z.object({
  id: z.number(),
  applicationDate: z.string(),
  isGeneratedByAi: z.boolean(),
  name: z.string(),
  statementImageUrl: z.string(),
  statementImageLegend: z.string(),
  statement: z.string(),
  quantityOfCorrectAnswers: z.number(),
  quantityOfWrongAnswers: z.number(),
  quantityOfTries: z.number(),
  explanation: z.string(),
  videoExplanationUrl: z.string(),
  questionAlternativeList: z.array(questionAlternativeSchema),
  questionDifficultyLevel: questionDifficultyLevelSchema,
  subject: subjectSchema,
  subjectTopicList: z.array(subjectTopicSchema),
});

export const questionnaireSchema = z.object({
  id: z.number(),
  name: z.string(),
  viewCounter: z.number(),
  questionList: z.array(questionSchema),
  isPublic: z.boolean(),
  student: studentSchema,
});

export type QuestionnaireList = z.infer<typeof questionnaireSchema>;